* add basic support for elf * add basic coloration

+++ src/gadget.cpp
@@ -1,6 +1,6 @@
 #include "gadget.hpp"
 
-Gadget::Gadget(const char* ins, const char* ops, unsigned int size, unsigned int offset)
+Gadget::Gadget(const char* ins, const char* ops, unsigned int size, unsigned long long offset)
 : m_disassembly(ins), m_opcodes(ops), m_size(size), m_offset(offset)
 {
 }
@@ -14,7 +14,7 @@ const char* Gadget::get_disassembly(void) const
     return m_disassembly;
 }
 
-unsigned int Gadget::get_offset(void) const
+unsigned long long Gadget::get_offset(void) const
 {
     return m_offset;
 }
