+++ src/inc/gadget.hpp
@@ -4,12 +4,12 @@
 class Gadget
 {
     public:
-        explicit Gadget(const char* ins, const char* ops, unsigned int size, unsigned int offset = 0);
+        explicit Gadget(const char* ins, const char* ops, unsigned int size, unsigned long long offset = 0);
         ~Gadget(void);
 
         const char* get_disassembly(void) const;
 
-        unsigned int get_offset(void) const;
+        unsigned long long get_offset(void) const;
         unsigned int get_size(void) const;
         const char* get_opcodes(void) const;
 
@@ -18,7 +18,7 @@ class Gadget
         const char *m_opcodes;
 
         unsigned int m_size;
-        unsigned int m_offset;
+        unsigned long long m_offset;
 };
 
 #endif
